Load plugins relative to rbenv executable
This commit adds a search for plugins in ${bin_path}/../plugins after ${RBENV_ROOT}/plugins.

Pre-commit, rbenv only searched for plugins in:

  $RBENV_ROOT/plugins

In a multi-user installation, where rbenv is available system-wide, but $RBENV_ROOT (i.e. the Ruby install target) is user-local, only the user's plugin directory is searched.

This commit adds a second search of the plugin directory relative to the rbenv executable.  It uses the same method as is used for finding shell completions in libexec/rbenv-init.

  bin_path="$(abs_dirname "$0")"
  ${bin_path}/../plugins/*/bin

Plugin paths are appended to $bin_path in the order that they are found.  If a plugin exists in both the user-local and system-wide directories, then the user-local plugin will be used because it occurs first in the path list.

A corresponding addition is made to $RBENV_HOOK_PATH as well.